Migrate get-api-url module to TS
